refactor: replace directory and repositories
