+++ src/main/kotlin/jsonweb/exitserver/config/WebConfig.kt
@@ -1,4 +1,4 @@
-package jsonweb.exitserver.common
+package jsonweb.exitserver.config
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.CorsRegistry
