--- src/main/kotlin/jsonweb/exitserver/domain/cafe/CafeKeywordRepository.kt
@@ -1,4 +1,0 @@
-package jsonweb.exitserver.domain.cafe
-
-interface CafeKeywordRepository {
-}
