--- src/main/kotlin/jsonweb/exitserver/domain/theme/GenreRepository.kt
@@ -1,8 +1,0 @@
-package jsonweb.exitserver.domain.theme
-
-import org.springframework.data.jpa.repository.JpaRepository
-import java.util.*
-
-interface GenreRepository: JpaRepository<Genre, Long> {
-    fun findGenreByGenreName(genreName: String): Optional<Genre>
-}
