--- src/main/kotlin/jsonweb/exitserver/domain/theme/ThemeGenreRepository.kt
@@ -1,6 +1,0 @@
-package jsonweb.exitserver.domain.theme
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ThemeGenreRepository: JpaRepository<ThemeGenre, Long> {
-}
