+++ src/test/kotlin/jsonweb/exitserver/domain/cafe/CafeLikeRepositoryTest.kt
@@ -4,7 +4,7 @@ import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
 import jsonweb.exitserver.domain.cafe.entity.CafeLike
 import jsonweb.exitserver.domain.cafe.entity.UserAndCafe
-import jsonweb.exitserver.util.TestConfig
+import jsonweb.exitserver.config.TestConfig
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.jdbc.EmbeddedDatabaseConnection
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
