+++ src/test/kotlin/jsonweb/exitserver/domain/theme/ThemeServiceTest.kt
@@ -2,11 +2,9 @@ package jsonweb.exitserver.domain.theme
 
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
-import io.kotest.mpp.log
 import jsonweb.exitserver.domain.cafe.CafeRepository
 import jsonweb.exitserver.domain.cafe.entity.Cafe
-import jsonweb.exitserver.util.ModelMapperConfig
-import jsonweb.exitserver.util.TestConfig
+import jsonweb.exitserver.config.TestConfig
 
 import org.junit.jupiter.api.Assertions.*
 import org.modelmapper.ModelMapper
@@ -14,7 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.jdbc.EmbeddedDatabaseConnection
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.context.annotation.Import
 import org.springframework.test.context.TestPropertySource
